+++ register_types.cpp
@@ -1,6 +1,7 @@
 #include "register_types.h"
 #include "edition/voxel_tool.h"
 #include "edition/voxel_tool_buffer.h"
+#include "edition/voxel_tool_lod_terrain.h"
 #include "edition/voxel_tool_terrain.h"
 #include "editor/editor_plugin.h"
 #include "editor/fast_noise_lite/fast_noise_lite_editor_plugin.h"
@@ -100,6 +101,7 @@ void register_voxel_types() {
 	ClassDB::register_class<VoxelRaycastResult>();
 	ClassDB::register_class<VoxelTool>();
 	ClassDB::register_class<VoxelToolTerrain>();
+	ClassDB::register_class<VoxelToolLodTerrain>();
 	// I had to bind this one despite it being useless as-is because otherwise Godot lazily initializes its class.
 	// And this can happen in a thread, causing crashes due to the concurrent access
 	ClassDB::register_class<VoxelToolBuffer>();